Top-of-the-day point timing & terramon config

+++ velodyne_pointcloud/include/velodyne_pointcloud/pointcloudXYZIRT.hpp
@@ -65,7 +65,7 @@ public:
 private:
   sensor_msgs::PointCloud2Iterator<float> iter_x_, iter_y_, iter_z_, iter_intensity_;
   sensor_msgs::PointCloud2Iterator<uint16_t> iter_ring_;
-  sensor_msgs::PointCloud2Iterator<float> iter_time_;
+  sensor_msgs::PointCloud2Iterator<double> iter_time_;
 };
 }  // namespace velodyne_pointcloud
